Fix compilation on MacOS / GHC 8.10.7
Fixes errors along the lines of:
dear-imgui  > [ 2 of 17] Compiling DearImGui.GLFW
dear-imgui  > error: unknown type name 'constexpr'

So the "-std=c++11" option is not being passed to the C++ compiler on
MacOS.
The issue seems related to https://github.com/haskell/cabal/issues/6421
Doesn't seem like this should be necessary with newer version of Cabal,
but nevertheless, this fix 'works for me'.

May be worth revisiting in the future.